Update jsonlib.cpp
I too am seeing this issue when using the alpha preview version of the ESP arduino core. Dropping back to the 1.06 release compiles correctly. I find that the simple strings test example compiles correctly with the alpha (maybe because it does not call std::string::npos ?) At any rate the issue seems to lie with the alpha version of the core.

For now, I work around the issue by editing src/jsonlib.cpp line 76 to avoid the issue:
```
  if (json.indexOf(name) == name.length()) return json.substring(0,0);
```